refactored: new ui.js and userAction.js
